Time: 0 ms (100%), Space: 47.3 MB (49.04%) - LeetHub

+++ 0268-missing-number/0268-missing-number.java
@@ -1,11 +1,11 @@
 class Solution {
     public int missingNumber(int[] nums) {
-        int arsum =0;
-        int n = nums.length;
-       int nsum = n*(n+1)/2;
+        long arsum =0;
+        long n = nums.length;
+       long nsum = n*(n+1)/2;
         for(int i=0; i< n; i++){
             arsum += nums[i];
         }
-        return nsum-arsum;
+        return (int)(nsum - arsum);
     }
 }
